Fixed copy

+++ NukeBuildHelpers/Common/FileExtensions.cs
@@ -6,15 +6,23 @@ internal static class AbsolutePathExtensions
 {
     public static void CopyFilesRecursively(this AbsolutePath path, AbsolutePath targetPath)
     {
-        foreach (string dirPath in Directory.GetDirectories(path.ToString(), "*", SearchOption.AllDirectories))
+        if (path.FileExists())
         {
-            Directory.CreateDirectory(dirPath.Replace(path.ToString(), targetPath));
+            Directory.CreateDirectory(targetPath.Parent);
+            File.Copy(path.ToString(), targetPath.ToString(), true);
         }
-
-        foreach (string newPath in Directory.GetFiles(path.ToString(), "*.*", SearchOption.AllDirectories))
+        else
         {
-            Directory.CreateDirectory(AbsolutePath.Create(newPath.Replace(path.ToString(), targetPath.ToString())).Parent);
-            File.Copy(newPath, newPath.Replace(path.ToString(), targetPath.ToString()), true);
+            foreach (string dirPath in Directory.GetDirectories(path.ToString(), "*", SearchOption.AllDirectories))
+            {
+                Directory.CreateDirectory(dirPath.Replace(path.ToString(), targetPath));
+            }
+
+            foreach (string newPath in Directory.GetFiles(path.ToString(), "*.*", SearchOption.AllDirectories))
+            {
+                Directory.CreateDirectory(AbsolutePath.Create(newPath.Replace(path.ToString(), targetPath.ToString())).Parent);
+                File.Copy(newPath, newPath.Replace(path.ToString(), targetPath.ToString()), true);
+            }
         }
     }
